requirements.txt
flask
python-dotenv
twilio
requests
cryptography
google-generativeai
sentence-transformers
chromadb
